Refactor inventory and order detail handling: change quantity type to BigDecimal, update related methods, and enhance validation

+++ src/main/java/hoanght/posapi/service/InventoryService.java
@@ -1,5 +1,7 @@
 package hoanght.posapi.service;
 
+import java.math.BigDecimal;
+
 public interface InventoryService {
-    void adjustInventory(Long productId, Long quantityChange);
+    void adjustInventory(Long productId, BigDecimal quantityChange);
 }
